tests: move ex-Cucumber steps to `FeatureHelpers::Steps`
