--- features/step_definitions/comments_steps.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-def i_write_a_new_comment body
-  first("#comment_body").set body
-end
