+++ spec/support/feature_helpers/step_helpers.rb
@@ -3,11 +3,12 @@
 require_relative '../../../features/support/cucumber_helpers/paths'
 require_relative '../../../features/support/cucumber_helpers/selectors'
 
-Dir[Rails.root.join("features/step_definitions/**/*.rb")].each { |f| require f }
+Dir[Rails.root.join("spec/support/feature_helpers/steps/**/*.rb")].each { |f| require f }
 
 module FeatureHelpers
   module StepHelpers
     include CucumberHelpers::Paths
     include CucumberHelpers::Selectors
+    include FeatureHelpers::Steps
   end
 end
