+++ spec/support/feature_helpers/steps/user_steps.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+module FeatureHelpers
+  module Steps
+    def login_programmatically user
+      login_as user, scope: :user, run_callbacks: false
+    end
+
+    def i_log_in_as name
+      user = User.find_by!(name: name)
+      login_programmatically user
+    end
+
+    def i_log_in_as_a_user_named name
+      user = create :user, name: name
+      login_programmatically user
+    end
+
+    def i_log_in_as_a_catalog_editor
+      user = create :user, :editor
+      login_programmatically user
+    end
+
+    def i_log_in_as_a_catalog_editor_named name
+      user = create :user, :editor, name: name
+      login_programmatically user
+    end
+
+    def i_log_in_as_a_superadmin_named name
+      user = create :user, :editor, :superadmin, name: name
+      login_programmatically user
+    end
+  end
+end
